tests/wpt/tests/css/css-inline/parsing/vertical-align-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Inline Layout: parsing vertical-align with valid values</title>
<link rel="help" href="https://www.w3.org/TR/CSS2/visudet.html#propdef-vertical-align">
<link rel="help" href="https://drafts.csswg.org/css-inline/#propdef-vertical-align">
<meta name="assert" content="vertical-align only supports the spec grammar.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("vertical-align", "none");

// Two baseline-source values
test_invalid_value("vertical-align", "first last");
test_invalid_value("vertical-align", "first, last");

// Two baseline-shift values
test_invalid_value("vertical-align", "sub super");
test_invalid_value("vertical-align", "super 10px");
test_invalid_value("vertical-align", "20% sub");

// Two alignment-baseline values
test_invalid_value("vertical-align", "baseline middle");
test_invalid_value("vertical-align", "text-top, bottom");
</script>
</body>
</html>
